Improve physics loss and migrate pipeline to high-res dataset + Zarr
- switch dynamic physics term from laplacian(input) to laplacian(output)

- generate and use new higher-resolution dataset (new_detailed_heat_sim_f64)

- adapt training/data loading to generalized timestep/resolution handling

- add on-the-fly normalization from normalization_values.json

- add Zarr conversion + CPU slurm jobs for sim/preprocess/conversion

+++ configs/train_config.py
@@ -1,3 +1,3 @@
 import torch
 
-TRAIN_DTYPE = torch.float64
+TRAIN_DTYPE = torch.float32
